+++ src/Driver/RedisProxySetDriver.php
@@ -5,7 +5,9 @@ declare(strict_types=1);
 namespace Efabrica\HermesExtension\Driver;
 
 use Closure;
+use Efabrica\HermesExtension\Driver\Interfaces\ForkableDriverInterface;
 use Efabrica\HermesExtension\Driver\Interfaces\QueueAwareInterface;
+use Efabrica\HermesExtension\Driver\Traits\ForkableDriverTrait;
 use Efabrica\HermesExtension\Driver\Traits\ProcessSignalTrait;
 use Efabrica\HermesExtension\Driver\Traits\QueueAwareTrait;
 use Efabrica\HermesExtension\Heartbeat\HeartbeatBehavior;
@@ -23,7 +25,7 @@ use Tomaj\Hermes\MessageSerializer;
 use Tomaj\Hermes\SerializeException;
 use Tomaj\Hermes\Shutdown\ShutdownException;
 
-final class RedisProxySetDriver implements DriverInterface, QueueAwareInterface
+final class RedisProxySetDriver implements DriverInterface, QueueAwareInterface, ForkableDriverInterface
 {
     use MaxItemsTrait;
     use ShutdownTrait;
@@ -31,6 +33,7 @@ final class RedisProxySetDriver implements DriverInterface, QueueAwareInterface
     use HeartbeatBehavior;
     use QueueAwareTrait;
     use ProcessSignalTrait;
+    use ForkableDriverTrait;
 
     /** @var array<int, string>  */
     private array $queues = [];
@@ -101,7 +104,7 @@ final class RedisProxySetDriver implements DriverInterface, QueueAwareInterface
                 $messageString = null;
                 $foundPriority = null;
 
-                if ($this->canContinue()) {
+                if ($this->canContinue() && !$this->hasActiveChildFork()) {
                     $messageString = $this->pop($this->getKey($priority));
                     $foundPriority = $priority;
                 }
@@ -122,7 +125,11 @@ final class RedisProxySetDriver implements DriverInterface, QueueAwareInterface
                 }
                 $message = $this->serializer->unserialize($messageString);
                 $accessor->setMessage($message, $foundPriority);
-                $callback($message, $foundPriority);
+                $this->doForkProcess(
+                    static function () use ($callback, $message, $foundPriority) {
+                        $callback($message, $foundPriority);
+                    }
+                );
                 $accessor->clear();
                 $this->incrementProcessedItems();
             } elseif ($this->refreshInterval) {